Refactoring comopnents
