+++ frontend/src/lib/components/LoadingBlocker.svelte
